+++ projects/spotfire-wrapper/src/lib/spotfire-webplayer.ts
@@ -86,7 +86,7 @@ export class SpotfireParameters {
   domid: string;
   sid: string;
   customization: Customization;
-  version: string = '7.14'; // Nominal fall back value
+  version = '7.14'; // Nominal fall back value
   debug = false;
   reloadAnalysisInstance = false;
   document: Document;
@@ -280,6 +280,20 @@ export class SpotfireReporting {
   getReports = (callback) => this.exp.getReports(callback);
   /** Export the active visual as image. The image will be opened in a new browser tab or window. */
   exportActiveVisualAsImage = (width = 800, height = 600) => this.exp.exportActiveVisualAsImage(width, height);
+}
+
+/**
+ * @description
+ * An observable object that decribes the status of the TIBCO Spotfire Server.
+ */
+export class SpotfireServer {
+  isOnline: boolean;
+  serverUrl: string;
+  statusMessage: string;
+  constructor(serverUrl: string, isOnline: boolean) {
+    this.isOnline = isOnline;
+    this.serverUrl = serverUrl;
+  }
 }
 
 /**