+++ projects/spotfire-wrapper/src/lib/viewer/spotfire-viewer.component.ts
@@ -13,7 +13,8 @@ import { tap } from 'rxjs/operators';
 
 import { DocumentService } from '../document.service';
 import { SpotfireCustomization, SpotfireFilter } from '../spotfire-customization';
-import { Application, Document, DocMetadata, SpotfireFiltering, SpotfireParameters, SpotfireReporting } from '../spotfire-webplayer';
+import { SpotfireServerService } from '../spotfire-server.service';
+import { Application, Document, DocMetadata, SpotfireFiltering, SpotfireParameters, SpotfireReporting, SpotfireServer } from '../spotfire-webplayer';
 
 // https://community.tibco.com/wiki/tibco-spotfire-javascript-api-overview
 // https://community.tibco.com/wiki/mashup-example-multiple-views-using-tibco-spotfire-javascript-api
@@ -81,22 +82,6 @@ export class SpotfireViewerComponent implements OnChanges, OnInit {
    */
   @Input() customization: SpotfireCustomization | string;
 
-  /**
-   * @description
-   * Optional. Array of filters that will be applied once page is loaded.
-   */
-
-  @Input() set filters(value: Array<SpotfireFilter> | string) {
-    if (typeof value === 'string') {
-      const allFilters: Array<SpotfireFilter> = [];
-      JSON.parse(value).forEach((m: SpotfireFilter) => allFilters.push(new SpotfireFilter(m)));
-      this._filters = allFilters;
-    } else {
-      this._filters = value as Array<SpotfireFilter>;
-    }
-    this.setFilters();
-  }
-
   @Input() markingOn: {} | string;
   @Input() maxRows = 10;
 
@@ -141,8 +126,17 @@ export class SpotfireViewerComponent implements OnChanges, OnInit {
    */
   @Output() filtering: EventEmitter<SpotfireFiltering> = new EventEmitter();
 
+  /**
+   * @description
+   * Optional. Emit an instance of a SpotfireServer whenever TIBCO Spotfire Server status is available.
+   */
+  @Output() serverStatusEvent: EventEmitter<SpotfireServer> = new EventEmitter();
+
   view: any;
   longTime = false;
+
+  readonly filter$: Observable<Array<{}>>;
+  readonly marker$: Observable<{}>;
 
   protected spotParams: SpotfireParameters = new SpotfireParameters();
   private _filters: Array<SpotfireFilter>;
@@ -150,22 +144,41 @@ export class SpotfireViewerComponent implements OnChanges, OnInit {
   private app: Application;
   /* Filtering observables, emitter and subject*/
   private filterSubject = new BehaviorSubject<Array<{}>>([]);
-  // tslint:disable-next-line:member-ordering
-  public filter$: Observable<Array<{}>> = this.filterSubject.asObservable();
 
   /* Marking observables, emitter and subject*/
   private markerSubject = new BehaviorSubject<{}>({});
-  // tslint:disable-next-line:member-ordering
-  public marker$: Observable<{}> = this.markerSubject.asObservable();
   private markedRows = {};
 
-  constructor(public docSvc: DocumentService) {
+  constructor(public docSvc: DocumentService, private spotfireServerSvc: SpotfireServerService) {
+    this.filter$ = this.filterSubject.asObservable();
+    this.marker$ = this.markerSubject.asObservable();
     this.doConsole('Welcome to Wrapper for TIBCO Spotfire(R)!');
     setTimeout(() => this.longTime = true, 6000);
   }
 
+  /**
+   * @description
+   * Optional. Array of filters that will be applied once page is loaded.
+   */
+
+  @Input() set filters(value: Array<SpotfireFilter> | string) {
+    if (typeof value === 'string') {
+      const allFilters: Array<SpotfireFilter> = [];
+      JSON.parse(value).forEach((m: SpotfireFilter) => allFilters.push(new SpotfireFilter(m)));
+      this._filters = allFilters;
+    } else {
+      this._filters = value as Array<SpotfireFilter>;
+    }
+    this.setFilters();
+  }
+
   ngOnInit(): void {
-    if (!this.version) this.version = DEFAULT_VERSION;
+    this.spotfireServerSvc.monitorSpotfireServerStatus(this.url);
+    this.spotfireServerSvc.serverStatusEvent.subscribe((e: SpotfireServer) => {
+      this.doConsole('SPOTFIRE-SERVER-SERVICE received event ' + JSON.stringify(e));
+      this.serverStatusEvent.emit(e);
+    });
+    if (!this.version) { this.version = DEFAULT_VERSION; }
     this.doConsole('OnInit', this.url, this.path, this.version);
     this.display();
   }
@@ -197,7 +210,8 @@ export class SpotfireViewerComponent implements OnChanges, OnInit {
       this.markingOn = JSON.parse(this.markingOn);
     }
 
-    this.doConsole('display', changes, this.url, this.version, this.path, 'PAGE=', this.page, this.customization, this.maxRows, this.app, this.markingOn);
+    this.doConsole('display', changes, this.url, this.version, this.path, 'PAGE=', this.page, this.customization, this.maxRows,
+      this.app, this.markingOn);
     if (!changes || changes.url) {
       this.openWebPlayer(this.url, this.path, this.customization);
     } else if (this.app && changes.page) {
@@ -221,7 +235,7 @@ export class SpotfireViewerComponent implements OnChanges, OnInit {
    *
    * @param page the document page that will be displayed
    */
-  public openPage(page: string) {
+  openPage(page: string) {
     this.displayInfoMessage(`${this.url}/${this.path}/${page ? page : ''}...`);
     this.doConsole(`SpotfireViewerComponent openPage(${page})`);
     this.page = page;
@@ -267,6 +281,7 @@ export class SpotfireViewerComponent implements OnChanges, OnInit {
     this.docSvc.openWebPlayer$(this.spotParams).subscribe(
       doc => this.afterDisplay(doc),
       err => this.displayErrorMessage(err));
+
   }
 
   /**