+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+<a name="0.12.0"></a>
+## [0.12.0] (2020-08-16)
+
+### Features
+ * Added a new service that monitors the status of the TIBCO Spotfire Server.
+ * Add a new output parameter `serverStatusEvent` that emits a `SpotfireServer` class after the TIBCO Spotfire Server has been monitored.
+ * Demo1 can demonstrate the code.
+
 <a name="0.11.6"></a>
 ## [0.11.6] (2020-08-15)